requirements.txt
qdrant-client>=1.8.0
networkx>=3.0
litellm>=1.20.0
mcp>=1.2.0
pydantic>=2.5.0
fastembed>=0.2.0
python-dotenv>=1.0.0
rich>=13.7.0
